Split OpenAI entity
